Moved updater logic into sperate class
