+++ MyMoney.Local/Program.cs
@@ -1,10 +1,6 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
-using System.IO.Compression;
-using System.Linq;
-using System.Net.Http;
-using System.Net.Http.Headers;
 using System.Threading.Tasks;
 
 namespace MyMoney.Local
@@ -13,7 +9,9 @@ namespace MyMoney.Local
    {
       public static void Main(string[] args)
       {
-         Task.Run(async () => await EnsureLatestVersion()).Wait();
+         var updater = new Updater();
+
+         Task.Run(async () => await updater.EnsureLatestVersion()).Wait();
 
          Launch();
       }
@@ -23,107 +21,15 @@ namespace MyMoney.Local
          if (!File.Exists(Constants.WebAppPath))
             return;
 
-         try
+         Console.WriteLine("Starting MyMoney...");
+         var process = Process.Start(Constants.WebAppPath);
+         if (process == null)
          {
-            Console.WriteLine("Starting MyMoney...");
-            var process = Process.Start(Constants.WebAppPath);
-            if (process == null)
-            {
-               Console.WriteLine("Failed to start MyMoney");
-            }
-            else
-            {
-               process.WaitForExit();
-            }
+            Console.WriteLine("Failed to start MyMoney");
          }
-         catch (Exception e)
+         else
          {
-            Console.WriteLine(e);
-
-            Console.WriteLine("\n\nPress any key to close...");
-            Console.ReadLine();
-         }
-      }
-
-      private static async Task<string> GetLatestTag(HttpClient client)
-      {
-         Console.WriteLine("Searching for latest version...");
-         var response = await client.GetAsync($"{Constants.RepositoryURL}/releases/latest");
-         var redirect = response.RequestMessage.RequestUri;
-         return redirect.ToString().Split('/').Last();
-      }
-
-      private static bool ShouldDownloadLatest(Version latest)
-      {
-         // If the web app doesnt exist, download the latest
-         if (!File.Exists(Constants.WebAppPath))
-            return true;
-
-         // Check the local version
-         var fileData = FileVersionInfo.GetVersionInfo(Constants.WebAppPath);
-         var localVersion = new Version(fileData.FileVersion);
-         return latest > localVersion;
-      }
-
-      private static void EmptyExistingFolder()
-      {
-         if (Directory.Exists(Constants.WebAppFolder))
-         {
-            Directory.Delete(Constants.WebAppFolder, true);
-         }
-
-         Directory.CreateDirectory(Constants.WebAppFolder);
-      }
-
-      private static async Task DownloadLatest(HttpClient client, string latestTag, Version latestVersion)
-      {
-         var assetPath = string.Format(Constants.RepositoryURL + "/releases/download/{0}/MyMoney_win64_{1}-{2}-{3}.zip", latestTag, latestVersion.Major, latestVersion.Minor, latestVersion.Build);
-
-         var fileResponse = await client.GetAsync(assetPath);
-
-         var zipFilePath = Path.Combine(Directory.GetCurrentDirectory(), $"{DateTime.UtcNow.ToFileTimeUtc()}.zip");
-
-         await using (var ms = await fileResponse.Content.ReadAsStreamAsync())
-         {
-            await using (var fs = File.Create(zipFilePath))
-            {
-               ms.Seek(0, SeekOrigin.Begin);
-               ms.CopyTo(fs);
-            }
-         }
-
-         ZipFile.ExtractToDirectory(zipFilePath, Constants.WebAppFolder);
-
-         File.Delete(zipFilePath);
-      }
-
-      private static async Task EnsureLatestVersion()
-      {
-         using (var client = new HttpClient())
-         {
-            client.DefaultRequestHeaders.Accept.Clear();
-            client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/vnd.github.v3+json"));
-
-            try
-            {
-               var latestTag = await GetLatestTag(client);
-               var latestVersion = new Version(latestTag[1..]);
-
-               if (!ShouldDownloadLatest(latestVersion))
-               {
-                  Console.WriteLine($"Already using latest version {latestTag}");
-                  return;
-               }
-
-               Console.WriteLine($"Downloading latest version {latestTag}...");
-               EmptyExistingFolder();
-               await DownloadLatest(client, latestTag, latestVersion);
-            }
-            catch (Exception e)
-            {
-               Console.WriteLine("Failed check for new version");
-               Console.WriteLine(e);
-            }
+            process.WaitForExit();
          }
       }
    }
